+++ Hearthstone Deck Tracker/Controls/Overlay/Constructed/PlayerResourcesWidget/PlayerResourcesWidget.xaml.cs
@@ -1,0 +1,13 @@
+﻿using System.Windows.Controls;
+
+
+namespace Hearthstone_Deck_Tracker.Controls.Overlay.Constructed.PlayerResourcesWidget;
+
+public partial class PlayerResourcesWidget : UserControl
+{
+	public PlayerResourcesWidget()
+	{
+		InitializeComponent();
+	}
+}
+